pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "pycequeau"
version = "0.1.1"
description = "A library that obtains the necessary files to build and run the CEQUEAU hydrological-water temperature model"
authors = [{name = "Eisinhower Rincon", email = "eisinhower.rincon@inrs.ca"}]
readme = {file = "README.rst", content-type = "text/x-rst"}
license = {file = "LICENSE"}
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Programming Language :: Python :: 3.10",
    "Topic :: Scientific/Engineering :: Hydrology",
]
dependencies = [
    "pyproj",
    "xarray",
    "scipy",
    "matplotlib",
    "shapely",
    "geopandas",
    "rasterstats",
    "netcdf4",
    "pytest",
    "autopep8",
    "ipykernel",
    "rioxarray",
    "dask",
    "xarray-spatial",
]

[project.urls]
Homepage = "https://github.com/erinconv/pycequeau"
Repository = "https://github.com/erinconv/pycequeau"
Documentation = "https://pycequeau.readthedocs.io/"

[project.optional-dependencies]
dev = [
    "toml",
    "pytest",
    "autopep8",
    "ipykernel",
    # Sphinx documentation
    "sphinx",
    "sphinx-rtd-theme",
    "sphinx-copybutton",
    "nbsphinx",
    "sphinx-autodoc-typehints",
    "sphinx-mathjax-offline",
]

[tool.flit.sdist]
include = [
    "docs/",
    "*.rst",
    "*.yml",
    "*.yaml",
]
